feat: import missing usings
