+++ lua/easy-dotnet/constants.lua
@@ -5,7 +5,6 @@ M.sign_namespace = "EasyDotnetTestSignGroup"
 M.lsp_client_name = "easy_dotnet"
 M.lsp_html_client_name = "easy_dotnet_html"
 M.lsp_projx_client_name = "easy_dotnet_projx"
-M.lsp_in_process_client_name = "easy_dotnet_in_process"
 M.debug_adapter_name = "easy-dotnet"
 M.compat_command = "dotnet easydotnet compat"
 M.server_quickfix_title = "dotnet server quickfix"
